requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
batchgenerators==0.25
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.26.3
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
efficientnet-pytorch==0.7.1
einops==0.8.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.18
fastjsonschema==2.21.1
filelock==3.12.0
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
future==0.18.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==3.2.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
imageio==2.27.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.2.0
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazy_loader==0.2
linecache2==1.0.0
lit==16.0.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
munch==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.3.2
notebook==7.4.1
notebook_shim==0.2.4
numcodecs==0.11.0
numpy==1.24.2
nvflare==2.5.2
nvflare-nightly @ git+https://github.com/NVIDIA/NVFlare.git@872cdc06bff7dc4a52eba2ac72813c551fc2f796
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
opencv-python==4.7.0.72
overrides==7.7.0
packaging==23.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.3.7
pretrainedmodels==0.7.4
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyhocon==0.3.61
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.8.2
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.3
shellingham==1.5.4
simpleitk==2.5.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.16
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.18.1
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.6.12
tinycss2==1.4.0
tomli==2.2.1
torch==2.7.0
torchio==0.20.8
torchsummary==1.5.1
torchvision==0.15.1
tornado==6.4.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
triton==3.3.0
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
unittest2==1.1.0
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.20.0
zarr==2.14.1
